Moving DeploymentFiltering to its module.
